substruct match items get high level interface
